Change return type for vcpop and vfirst to uint and int
Using `long` introduces ambiguity to rely on the architecture specified
(rv32/rv64). This is unnecessary since it is likely to be impossible for
a vector length to exceed the value range of an `int`.

+++ rvv-intrinsic-generator/rvv_intrinsic_gen/templates/mask_template.py
@@ -71,7 +71,7 @@ def render(G, op_list, type_list, sew_list, lmul_list, decorator_list):
         G.func(
             inst_info_m,
             name="{OP}_m_b{MLEN}".format_map(args) + decorator.func_suffix,
-            return_type=type_helper.ulong,
+            return_type=type_helper.uint,
             **decorator.mask_args(type_helper.m),
             vs2=type_helper.m,
             vl=type_helper.size_t)
@@ -79,7 +79,7 @@ def render(G, op_list, type_list, sew_list, lmul_list, decorator_list):
         G.func(
             inst_info_m,
             name="{OP}_m_b{MLEN}".format_map(args) + decorator.func_suffix,
-            return_type=type_helper.long,
+            return_type=type_helper.int,
             **decorator.mask_args(type_helper.m),
             vs2=type_helper.m,
             vl=type_helper.size_t)
